requirement.txt
pip install 

faiss-cpu==1.7.4
sentence_transformers==3.0.1
groq==0.18.0
transformers==4.41.2
numpy==1.26.0